requirements.txt
#Part 1 dependisies
azure-ai-formrecognizer==3.3.3
python-dotenv==1.0.1
openai==1.65.4
streamlit==1.43.0
celery==5.4.0
redis==5.2.1
#Part 2 dependisies
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
beautifulsoup4==4.13.3
lxml==5.3.1
scikit-learn==1.6.1
numpy==2.2.3
faiss-cpu==1.10.0